Tested verify function and working.
Only modules that declared access can emit functions. It also means there is no hacky way to call the backbone from outside to a function inside of a module without making the module declare it. Should prevent some sloppy coding with any luck. (of course that's not guaranteed, and if this is a dumb design, it can be ammended)